[ET-VK] Introduce `vTensorPtr` to prevent reference invalidation and remove `get_val()` API
Pull Request resolved: https://github.com/pytorch/executorch/pull/2978

## Context

Currently when writing operators developers will save a reference to a `vTensor` retrieved from a `ComputeGraph`'s list of `values_` like so:

```
vTensor& vten = graph.get_val(vref).toTensor();
```

However, this is dangerous since if any values are added once the reference has been stored, `values_` which is a `std::vector` may have been resized and therefore have its contents moved, meaning the reference is now invalid.

To protect against this, this changeset introduces the `vTensorPtr` class which is a wrapper around a `vTensor*`.  When constructed, it will increment a counter in the `ComputeGraph` instance, and when destroyed it will decrement the counter. `ComputeGraph` cannot add any values while the counter is not zero.

Since `Value` can be converted to other non-trivial types, this changeset also removes the `get_val` function entirely to guard against unsafe behaviour.
ghstack-source-id: 222224052
@exported-using-ghexport

Differential Revision: [D55984187](https://our.internmc.facebook.com/intern/diff/D55984187/)