[ET-VK] Bump Vulkan API requirement to 1.1 and enable 16 bit and 8 bit types in buffer storage
Pull Request resolved: https://github.com/pytorch/executorch/pull/3058

## Context

Enable use of explicit fp16 and int8 types in GPU storage buffers via the following extensions:

* [VK_KHR_16bit_storage](https://registry.khronos.org/vulkan/specs/1.3-extensions/man/html/VK_KHR_16bit_storage.html)
* [VK_KHR_8bit_storage](https://registry.khronos.org/vulkan/specs/1.3-extensions/man/html/VK_KHR_8bit_storage.html)
* [VK_KHR_shader_float16_int8](https://registry.khronos.org/vulkan/specs/1.3-extensions/man/html/VK_KHR_shader_float16_int8.html)

The first two enables usage of 16-bit and 8-bit types in storage buffers, while the last one enables using those types in arithmetic operations.

By enabling these extensions and checking that the device supports the required features, explicit fp16 and int8 types can be used in compute shaders, as demonstrated by the added test.

Vulkan 1.1 is required in order to access `vkGetPhysicalDeviceFeatures2`, which is required to query whether the device support 16bit and 8bit types. This should be a fairly straightforward version bump as Vulkan 1.1 is supported by the vast majority of Android devices.
ghstack-source-id: 222727208
@exported-using-ghexport

Differential Revision: [D56164239](https://our.internmc.facebook.com/intern/diff/D56164239/)