Implemement smoke_test.py for pip wheel jobs
This script is run by CI after building the executorch wheel. Before
running this, the job will install the matching torch package as well as
the newly-built executorch package and its dependencies.

For now we test the export of a simple model, and try executing it using
the runtime pybindings.

Test Plan:
```
./install_requirements.sh
python build/packaging/smoke_test.py
```

ghstack-source-id: 4bacc007c069c29b1ce60fd4afac641a7a2f772a
Pull Request resolved: https://github.com/pytorch/executorch/pull/3474