Audit and update the pip package metadata
Fill out the recommended `project` keys, most of which will affect the
web page that PyPI will render for the `executorch` package.

See
https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#about-your-project
for the latest guidance.

Use
https://github.com/pytorch/pytorch/blob/a21327e0b03cc18850a0608be2d9c5bd38fd4646/setup.py#L1394
as a guide for the actual values.

Add a README-wheel.md file that will be included in the wheel, and will
become the main page contents on PyPI.

Test Plan:
* Installed the package with `./install_requirements.sh`
* Looked at the files under ~/miniconda3/envs/executorch/lib/python3.10/site-packages/executorch-0.2.0a0+1a499e0.dist-info. METADATA and LICENSE both contain the new metadata.

ghstack-source-id: 1c2a97e6d82b30c6c2b4eedb95182f740af84f53
Pull Request resolved: https://github.com/pytorch/executorch/pull/3476